adds command line mp3 writing options to midi_alignment

+++ experiments/midi_alignment.py
@@ -18,6 +18,11 @@ sys.path.append('../')
 import align_midi
 import scipy.io
 
+write_mp3 = False
+if len(sys.argv) > 1:
+  if sys.argv[1] == '-w':
+    write_mp3 = True
+
 # <codecell>
 
 SF2_PATH = '../../Performer Synchronization Measure/SGM-V2.01.sf2'
@@ -157,34 +162,35 @@ def align_one_file(mp3_filename, midi_filename, output_midi_filename, output_dia
     if output_diagnostics:
         # Save the figures
         plt.savefig(output_midi_filename.replace('.mid', '.pdf'))
-        # Load in the audio data (needed for writing out)
-        audio, fs = librosa.load(mp3_filename, sr=None)
-        # Synthesize the aligned midi
-        # midi_audio_aligned = m_aligned.fluidsynth()
-        midi_audio_aligned = m_aligned.fluidsynth(fs=fs, sf2_path=SF2_PATH)
+        if write_mp3:
+          # Load in the audio data (needed for writing out)
+          audio, fs = librosa.load(mp3_filename, sr=None)
+          # Synthesize the aligned midi
+          # midi_audio_aligned = m_aligned.fluidsynth()
+          midi_audio_aligned = m_aligned.fluidsynth(fs=fs, sf2_path=SF2_PATH)
 
-        # Trim to the same size as audio
-        if midi_audio_aligned.shape[0] > audio.shape[0]:
-            midi_audio_aligned = midi_audio_aligned[:audio.shape[0]]
-        else:
-            midi_audio_aligned = np.append(midi_audio_aligned, np.zeros(audio.shape[0] - midi_audio_aligned.shape[0]))
-        # Write out to temporary .wav file
-        librosa.output.write_wav(output_midi_filename.replace('.mid', '.wav'),
-                                 np.vstack([midi_audio_aligned, audio]).T, fs)
-        # Convert to mp3
-        subprocess.check_output(['ffmpeg',
-                         '-i',
-                         output_midi_filename.replace('.mid', '.wav'),
-                         '-ab',
-                         '128k',
-                         '-y',
-                         output_midi_filename.replace('.mid', '.mp3')])
-        # Remove temporary .wav file
-        os.remove(output_midi_filename.replace('.mid', '.wav'))
-        # Save a .mat of the results
-        scipy.io.savemat(output_midi_filename.replace('.mid', '.mat'),
-                         {'similarity_matrix': similarity_matrix,
-                          'p': p, 'q': q, 'score': score})
+          # Trim to the same size as audio
+          if midi_audio_aligned.shape[0] > audio.shape[0]:
+              midi_audio_aligned = midi_audio_aligned[:audio.shape[0]]
+          else:
+              midi_audio_aligned = np.append(midi_audio_aligned, np.zeros(audio.shape[0] - midi_audio_aligned.shape[0]))
+          # Write out to temporary .wav file
+          librosa.output.write_wav(output_midi_filename.replace('.mid', '.wav'),
+                                   np.vstack([midi_audio_aligned, audio]).T, fs)
+          # Convert to mp3
+          subprocess.check_output(['ffmpeg',
+                           '-i',
+                           output_midi_filename.replace('.mid', '.wav'),
+                           '-ab',
+                           '128k',
+                           '-y',
+                           output_midi_filename.replace('.mid', '.mp3')])
+          # Remove temporary .wav file
+          os.remove(output_midi_filename.replace('.mid', '.wav'))
+          # Save a .mat of the results
+          scipy.io.savemat(output_midi_filename.replace('.mid', '.mat'),
+                           {'similarity_matrix': similarity_matrix,
+                            'p': p, 'q': q, 'score': score})
     # If we aren't outputting a .pdf, show the plot
     else:
         plt.show()
